requirements.txt
langchain-text-splitters
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.0
